rename: 프로젝트 도메인 리패키징
