--- src/main/java/sixgaezzang/sidepeek/projects/domain/FileType.java
@@ -1,6 +1,0 @@
-package sixgaezzang.sidepeek.projects.domain;
-
-public enum FileType {
-    IMAGE,
-    VIDEO
-}
