+++ src/main/java/sixgaezzang/sidepeek/projects/domain/file/FileType.java
@@ -1,0 +1,6 @@
+package sixgaezzang.sidepeek.projects.domain.file;
+
+public enum FileType {
+    IMAGE,
+    VIDEO
+}
