+++ src/main/java/sixgaezzang/sidepeek/projects/domain/member/AuthorityType.java
@@ -1,4 +1,4 @@
-package sixgaezzang.sidepeek.projects.domain;
+package sixgaezzang.sidepeek.projects.domain.member;
 
 public enum AuthorityType {
     OWNER,
