+++ src/main/java/sixgaezzang/sidepeek/projects/domain/member/Member.java
@@ -1,4 +1,4 @@
-package sixgaezzang.sidepeek.projects.domain;
+package sixgaezzang.sidepeek.projects.domain.member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -15,6 +15,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import sixgaezzang.sidepeek.projects.domain.Project;
 import sixgaezzang.sidepeek.users.domain.User;
 
 @Entity
